Introduce new flags for native routing and encap. protocol
The long-term plan is replace the --tunnel flag with --tunnel-protocol
and --routing-mode flags. In some cases (e.g., egress gateway
today), we need to setup an ad-hoc tunnel for some of the traffic
while in native routing mode. In such a configuration, it is not
possible today to pick the tunneling protocol and we default to VXLAN.
By splitting --tunnel, we will allow users to specify the tunneling
protocol even while running in native routing mode.

This commit simply introduces the two new flags. We fatal if
--routing-mode is set and --tunnel isn't empty.

Signed-off-by: Paul Chaignon <paul@cilium.io>