etcdclient: use parent ctx and fix logic around it
[ upstream commit 94b0ee1ae775365be60a2fb52d2e200ed2d62ce6 ]

Previously when etcd client context was not set and client
created with client.New(*config) it would use context.TODO,
without any way to cancel the ctx.

In some cases (e.g. wrong endpoints configuration) it would
 be necessary to cancel the ctx, otherwise the client might
be stucked in a retry loop and in case of remoteConnectionController
there wouldn't be any way stop that.

This commit passes a parent context to the client itself.

This commit also fixes the closing of stopWatch channel and rest
of the resources when ctx is done/cancelled.

If ctx was closed and select statement choosed handling of that
case (<-ctx.Done()) the code would quit the for loop and handling
of <-w.stopWatch would never be possible which means calling
w.Stop() from outside (e.g. releaseOldConnection()) would hang
forever as the wg counter wouldn't be decremented.
Now when the context gets closed, it correctly closes all channels
and decrements the wg counter, just like when handled by
releaseOldConnection func which releases everything.

Signed-off-by: Ondrej Blazek <ondrej.blazek@firma.seznam.cz>
Signed-off-by: Nikolay Aleksandrov <nikolay@isovalent.com>