docs/egressgw: Explicitly mention lack of IPv6 support
[ upstream commit 9f583c2ef076b9de84eae3737b0ab946ba656626 ]

We mention that the egress gateway is for IPv4 traffic, but we never
explictly say that IPv6 isn't supported. This is apparently not obvious
to everyone so let's state it here.

This commit moves the feature incompatibilies subsection under the
Preliminary Considerations section. Another subsection will be added
there by the next commit.

Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Nikolay Aleksandrov <nikolay@isovalent.com>